Optional automated propagation of credentials to ETL process
By setting gdc.etl.send_credentials to true, the user can force
agent to send credentials to the remote ETL process.

If this feature is turned on, the target ETL can expect the
gdc_agent_zip_url param to include the URL of the uploaded
zip file including the HTTP Basic credentials. Similarly,
gdc_agent_manifest_url can be used for a manifest based upload.

The names of these ETL parameters can be configured in the
properties file as documented in the README.md file.